start to implement like state management start
Here, a postWithLikeStatus eneity was created. And relevant method updates done in post controller, service and repository. Take the user id as well when running getAllPosts api call is run so the data to display to user if they have already liked is available and can be used in state management when the post components are displayed in feed or elsewhere