resolve comment display and count issues
Issues with the comment display was a type mismatch in the Mongoose query. postId was passed as a string while MongoDB was expecting an ObjectId. That issue was resolved. relevant logging was also added to backend controller, service and repository implementation files.          As for the comment count issue, it was that the comment count was not incrementing when a new comment was added. This meant that the old count was still displayed in the post. As a temporary frontend fix, the add comment handler increments the displayed count by 1 when a comment is added. The actual new comment count value will be displayed when the feed is refreshed